Adds utility function for showing web view
