+++ v3/src/components/tool-shelf/help-button.tsx
@@ -3,10 +3,8 @@ import { Menu, MenuButton, MenuItem, MenuList } from "@chakra-ui/react"
 import HelpIcon from "../../assets/icons/icon-help.svg"
 import { useDocumentContent } from "../../hooks/use-document-content"
 import { getDefaultLanguage, t } from "../../utilities/translation/translate"
-import { kWebViewTileType } from "../web-view/web-view-defs"
-import { isWebViewModel } from "../web-view/web-view-model"
 import { kRightButtonBackground, ToolShelfButtonTag } from "./tool-shelf-button"
-import { logMessageWithReplacement } from "../../lib/log-message"
+import { showWebView } from "./tool-shelf-utilities"
 
 import "./tool-shelf.scss"
 
@@ -28,14 +26,7 @@ export const HelpShelfButton = () => {
     const url = helpURLs[locale]
 
     if (url !== undefined) {
-      documentContent?.applyModelChange(() => {
-        const tile = documentContent?.createOrShowTile?.(kWebViewTileType)
-        isWebViewModel(tile?.content) && tile?.content.setUrl(`${url}`)
-      }, {
-        undoStringKey: "V3.Undo.webView.show",
-        redoStringKey: "V3.Redo.webView.show",
-        log: logMessageWithReplacement("Show web view: %@", {url}, "document")
-      })
+      showWebView(url, documentContent)
     } else {
       window.open(showHelpURL, '_blank')
     }
